use apiv3 in /attachment/add
